requirements.txt
Django>=4.2,<5.0
djangorestframework
psycopg2-binary  # For PostgreSQL, remove if using SQLite
gunicorn  # For production server
django-import-export